requirements.txt
art==6.0
discord.py==2.3.1
playwright==1.36.0
requests==2.31.0
rich==13.4.2
